* commit '9f953751034a3813f6e30d67ba11b0cb9f28fea5':
  SSBD201903-106 Bugfix for static files

+++ src/main/java/pl/lodz/p/it/ssbd2019/ssbd03/web/servlets/StaticFilesServlet.java
@@ -21,9 +21,13 @@ public class StaticFilesServlet extends HttpServlet {
         String filename = URLDecoder.decode(request.getPathInfo().substring(1), "UTF-8");
         String path =  getServletContext().getRealPath("/static");
         File file = new File(path, filename);
-        response.setHeader("Content-Type", getServletContext().getMimeType(filename));
-        response.setHeader("Content-Length", String.valueOf(file.length()));
-        response.setHeader("Content-Disposition", "inline; filename=\"" + file.getName() + "\"");
-        Files.copy(file.toPath(), response.getOutputStream());
+        if(!file.isDirectory() && file.exists()) {
+            response.setHeader("Content-Type", getServletContext().getMimeType(filename));
+            response.setHeader("Content-Length", String.valueOf(file.length()));
+            response.setHeader("Content-Disposition", "inline; filename=\"" + file.getName() + "\"");
+            Files.copy(file.toPath(), response.getOutputStream());
+        } else {
+            response.setStatus(404);
+        }
     }
 }
